fix: persist Polymarket edge outputs + fix risk_debate merge conflict
Fix 4 — Polymarket edge persistence:
- Add polymarket_edges table to DB schema with cycle/city/market_slug/edge/kelly columns
- edge_detection.py: _persist_edge() writes to DB via DATABASE_URL on every compute
- orchestrator_v2.py: surface edge/kelly_fraction/opportunity metrics in
  ExecuteStepResponse for EDGE_DETECTION steps

Merge fix — resolve risk_debate name conflict from eloquent-feynman merge:
- Rename signals-based debate method to signal_debate() to avoid collision
  with the V-TR4 multi-persona risk_debate() (RiskDebateConsensus return type)
- victoria_node.py DebateGate action now calls signal_debate()
- test_accuracy_fixes.py updated to match new name; 37/37 tests pass

+++ internal/core/full_cycle.go
@@ -120,7 +120,7 @@ func (o *Orchestrator) RunFullCycle(ctx context.Context, cfg FullCycleConfig) (*
 	// Create a parent span for the entire orchestration cycle.
 	tracer := telemetry.Tracer()
 	spanCtx, cycleSpan := tracer.Start(ctx, telemetry.SpanOrchestratorCycle,
-		trace.WithAttributes(telemetry.AttrCycleID.Int64(int64(cycleID))),
+		trace.WithAttributes(telemetry.AttrCycleID.Int64(cycleID)),
 	)
 	defer cycleSpan.End()
 
@@ -228,7 +228,7 @@ func (o *Orchestrator) RunFullCycle(ctx context.Context, cfg FullCycleConfig) (*
 		nodeCtx, nodeSpan := tracer.Start(nodeCtx, telemetry.SpanNodeExecution,
 			trace.WithAttributes(
 				telemetry.AttrNodeName.String(nodeID),
-				telemetry.AttrCycleID.Int64(int64(cycleID)),
+				telemetry.AttrCycleID.Int64(cycleID),
 			),
 		)
 		execErr := entry.executor.Execute(nodeCtx)
